test: add test macros and auxiliary functions
This commit adds some test macros to help dealing with
some of the async functions from Javascript.
Furthermore it adds a new example to the mantine-test testing suite
to show an instance of a synchronous test.
Tests that "just render components" and verifies the rendered screen
can be done synchronously.
Components that have side effect, states and user interactions, on the
other hand have to be tested asynchronously.
Please, note the screen function instead of the screen variable from
react testing library. Although it has the exact same effect, that's been
done to work around the js/document being hooked only after the testing
library is loaded.